moved vis code
